+++ topiary-queries/queries/menhir/formatting.scm
@@ -1,0 +1,78 @@
+[
+  (comment)
+  (ocaml_comment)
+] @multi_line_indent_all @allow_blank_line_before @prepend_input_softline
+
+(
+  [
+    (comment)
+    (ocaml_comment)
+  ] @append_input_softline
+)
+
+(line_comment) @append_hardline
+
+(declaration) @prepend_hardline @allow_blank_line_before
+
+(declaration
+  [
+    "%token"
+    "%type"
+    "%start"
+    "%inline"
+    "%left"
+    "%right"
+    "%nonassoc"
+    "%parameter"
+  ] @append_space
+)
+
+(header
+  "%{" @append_hardline
+  (ocaml) @multi_line_indent_all
+  "%}" @prepend_hardline
+) @append_hardline @allow_blank_line_before
+
+(type
+  "<"
+  (ocaml_type)
+  ">" @append_space
+)
+
+(terminal_alias_attrs) @append_space
+(non_terminal) @append_space
+(strict_actual) @append_space
+
+(old_rule) @append_hardline @allow_blank_line_before
+
+(flags) @append_space
+
+(old_rule
+  ":" @append_input_softline
+)
+
+(old_rule
+  "|" @prepend_input_softline @append_space
+)
+
+(production_group
+  "|" @prepend_input_softline @append_space
+)
+
+(producer
+  (lid) @append_space
+  "=" @append_space
+)
+
+(producer) @append_space
+(producer ";" @prepend_antispace)
+
+(actual
+  "," @append_space
+)
+
+(action
+  "{" @append_space
+  (ocaml) @multi_line_indent_all
+  "}"
+)